+++ fire_seq_search_server/src/http_client/mod.rs
@@ -1,0 +1,1 @@
+pub mod endpoints;
