+++ fire_seq_search_server/src/lib.rs
@@ -2,11 +2,13 @@ pub mod post_query;
 pub mod load_notes;
 pub mod markdown_parser;
 mod language_detect;
+pub mod http_client;
 
 
 use log::{debug, info};
 use crate::post_query::highlight_keywords_in_body;
 use serde::Serialize;
+
 
 #[macro_use]
 extern crate lazy_static;
@@ -213,6 +215,45 @@ pub fn tokens_to_text_vec(tokens: &Vec<tantivy::tokenizer::Token>) -> Vec<String
     }
     token_text
 }
+
+
+
+pub fn decode_cjk_str(original: String) -> Vec<String> {
+    use urlencoding::decode;
+
+    let mut result = Vec::new();
+    for s in original.split(' ') {
+        let t = decode(s).expect("UTF-8");
+        debug!("Decode {}  ->   {}", s, t);
+        result.push(String::from(t));
+    }
+
+    result
+}
+
+use rayon::prelude::*;
+fn post_query_wrapper(top_docs: Vec<(f32, tantivy::DocAddress)>,
+                      term: &String,
+                      searcher: &tantivy::LeasedItem<tantivy::Searcher>,
+                      server_info: &ServerInformation) -> Vec<String> {
+    let term_tokens = tokenize_default(&term);
+    info!("get term tokens {:?}", &term_tokens);
+    // let mut result;
+    let result: Vec<String> = top_docs.par_iter()
+        .map(|&x| FireSeqSearchHitParsed::from_tantivy
+            (&searcher.doc(x.1).unwrap(),
+             x.0,
+             &term_tokens,
+             server_info)
+        )
+        // .map(|x| FireSeqSearchHitParsed::from_hit(&x))
+        .map(|p| serde_json::to_string(&p).unwrap())
+        .collect();
+    result
+}
+
+
+
 // ============= BELOW IS TEST CASES ====================
